% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functionsdataframe.R
\name{task10_df_map_peaks_to_genes}
\alias{task10_df_map_peaks_to_genes}
\title{Task 10 (data.frame): ATAC peaks → genes mapping}
\usage{
task10_df_map_peaks_to_genes()
}
\value{
Top 20 genes by total overlap.
}
\description{
Task 10 (data.frame): ATAC peaks → genes mapping
}
