% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functionsdatatable.R
\name{task10_map_peaks_to_genes}
\alias{task10_map_peaks_to_genes}
\title{Task 10 — Map ATAC peaks to genes}
\usage{
task10_map_peaks_to_genes()
}
\value{
Top 20 genes by total overlap length.
}
\description{
Task 10 — Map ATAC peaks to genes
}
