% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functionsdataframe.R
\name{task11_df_map_snps}
\alias{task11_df_map_snps}
\title{Task 11 (data.frame): SNPs → genes mapping}
\usage{
task11_df_map_snps()
}
\value{
Data.frame of high-impact variants per gene and sample.
}
\description{
Task 11 (data.frame): SNPs → genes mapping
}
