% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functionsdatatable.R
\name{task11_map_snps}
\alias{task11_map_snps}
\title{Task 11 — Map variants to genes}
\usage{
task11_map_snps()
}
\value{
List of variant counts and genes with high-impact variants.
}
\description{
Task 11 — Map variants to genes
}
