% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functionsdatatable.R
\name{task12_combine_cohorts}
\alias{task12_combine_cohorts}
\title{Task 12 — Combine cohorts}
\usage{
task12_combine_cohorts()
}
\value{
Summary of top variable genes by cohort and condition.
}
\description{
Task 12 — Combine cohorts
}
