% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functionsdataframe.R
\name{task12_df_combine_cohorts}
\alias{task12_df_combine_cohorts}
\title{Task 12 (data.frame): Combine cohorts and summarize}
\usage{
task12_df_combine_cohorts()
}
\value{
Summary of top 100 variable genes per cohort/condition.
}
\description{
Task 12 (data.frame): Combine cohorts and summarize
}
