% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functionsdataframe.R
\name{task1_df_filter_counts}
\alias{task1_df_filter_counts}
\title{Task 1 (data.frame): Filter, summarize, and group bulk counts}
\usage{
task1_df_filter_counts()
}
\value{
A list with per-gene and per-condition summaries.
}
\description{
Task 1 (data.frame): Filter, summarize, and group bulk counts
}
