% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functionsdataframe.R
\name{task2_df_add_qc_columns}
\alias{task2_df_add_qc_columns}
\title{Task 2 (data.frame): Add QC-style derived columns}
\usage{
task2_df_add_qc_columns()
}
\value{
Data.frame with log2 and high flag columns.
}
\description{
Task 2 (data.frame): Add QC-style derived columns
}
