% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functionsdataframe.R
\name{task3_df_join_metadata}
\alias{task3_df_join_metadata}
\title{Task 3 (data.frame): Join metadata and counts}
\usage{
task3_df_join_metadata()
}
\value{
Merged data.frame of counts and metadata.
}
\description{
Task 3 (data.frame): Join metadata and counts
}
