% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functionsdataframe.R
\name{task4_df_annotate_counts}
\alias{task4_df_annotate_counts}
\title{Task 4 (data.frame): Annotate counts with sample/patient info}
\usage{
task4_df_annotate_counts()
}
\value{
A list of per-patient totals and top genes per condition.
}
\description{
Task 4 (data.frame): Annotate counts with sample/patient info
}
