% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functionsdataframe.R
\name{task5_df_classify_labs}
\alias{task5_df_classify_labs}
\title{Task 5 (data.frame): Classify labs as normal or out_of_range}
\usage{
task5_df_classify_labs()
}
\value{
A list with abnormal rates by patient and lab.
}
\description{
Task 5 (data.frame): Classify labs as normal or out_of_range
}
