% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functionsdataframe.R
\name{task7_df_filter_genomic_peaks}
\alias{task7_df_filter_genomic_peaks}
\title{Task 7 (data.frame): Slice genomics windows}
\usage{
task7_df_filter_genomic_peaks()
}
\value{
Top 50 peaks in chr2 between 2–4 Mb.
}
\description{
Task 7 (data.frame): Slice genomics windows
}
