% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functionsdatatable.R
\name{task7_filter_genomic_peaks}
\alias{task7_filter_genomic_peaks}
\title{Task 7 — Filter genomic peaks}
\usage{
task7_filter_genomic_peaks()
}
\value{
Top 50 peaks on chr2 between 2–4 Mb by score.
}
\description{
Task 7 — Filter genomic peaks
}
