% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functionsdataframe.R
\name{task8_df_summarize_gene_stats}
\alias{task8_df_summarize_gene_stats}
\title{Task 8 (data.frame): Multi-column stats per group}
\usage{
task8_df_summarize_gene_stats()
}
\value{
Data.frame of summary stats for selected genes.
}
\description{
Task 8 (data.frame): Multi-column stats per group
}
