% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functionsdatatable.R
\name{task8_summarize_gene_stats}
\alias{task8_summarize_gene_stats}
\title{Task 8 — Summarize gene stats}
\usage{
task8_summarize_gene_stats()
}
\value{
Data.table of summary statistics for genes.
}
\description{
Task 8 — Summarize gene stats
}
