% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functionsdataframe.R
\name{task9_df_reshape_counts}
\alias{task9_df_reshape_counts}
\title{Task 9 (data.frame): Wide → long → wide reshape}
\usage{
task9_df_reshape_counts()
}
\value{
Data.frame reshaped for plotting.
}
\description{
Task 9 (data.frame): Wide → long → wide reshape
}
