% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functionsdatatable.R
\name{task9_reshape_counts}
\alias{task9_reshape_counts}
\title{Task 9 — Reshape counts wide→long→wide}
\usage{
task9_reshape_counts()
}
\value{
Data.table reshaped for plotting.
}
\description{
Task 9 — Reshape counts wide→long→wide
}
